refactor: extract case-lifecycle + password routes into routes/ (router split, final)
Pure structural move out of createApp; no handler logic changed. server.ts now holds zero
route registrations (thin wiring) — only the case-lock gate app.use and the terminal
error-handler app.use remain.

- routes/casePassword.ts: set/change/clear password, unlock, lock, lock-status.
- routes/caseLifecycle.ts: the case-core catch-all — list/create/status/archive/restore/
  delete, state/stats, encrypted export+import, ioc-blocklist, backups, integration pushes
  (iris/timesketch/misp/notion/clickup) + their status/reconnect, manual events, importer
  CRUD, jobs, activity-log, settings/env, setup/status, and the static app shell
  (/, /dashboard, /mobile, manifest, sw.js — moved out of startServer so server.ts has
  zero literal route registrations).

Graduated to RouteContext (shared with staying code, so graduated not moved): instanceSecret
(unlock-cookie secret, also used by the staying case-lock gate + readUnlockState),
runStateExclusive (per-case state mutex), ensureDropFolders (drop watcher shares it),
reloadImporters + importerPrecedence()/setImporterPrecedence (the staying detection seam
reads the same registry+precedence), rebuildTimesketchClient (mirrors rebuildIrisClient).
